refactor(ios/presentation): split ContentView access subcomponents
